pyproject.toml
# Rename to dev version
# This is dev version
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "automation_ide_dev"
version = "1.0.3"
authors = [
    { name = "JE-Chen", email = "jechenmailman@gmail.com" },
]
description = "AutomationEditor for multi automation"
requires-python = ">=3.10"
license-files = ["LICENSE"]
dependencies = [
    "je_editor_dev", "je_auto_control", "je_web_runner",
    "je_load_density", "je_api_testka", "je-mail-thunder",
    "automation-file", "PySide6==6.10.2", "test_pioneer", "paramiko"
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Win32 (MS Windows)",
    "Environment :: MacOS X",
    "Environment :: X11 Applications",
    "Operating System :: OS Independent"
]

[project.urls]
Homepage = "https://github.com/Intergration-Automation-Testing/AutomationEditor"
Code = "https://github.com/Intergration-Automation-Testing/AutomationEditor"

[project.readme]
file = "README.md"
content-type = "text/markdown"


[tool.setuptools.packages]
find = { namespaces = false }
